requirements.txt
pandas>=1.3.0
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0  # Optional for environment variables